% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predict10}
\alias{predict10}
\title{Prediction with Top Ten Features}
\usage{
predict10(X, b)
}
\arguments{
\item{X}{an [n x 10] matrix containing n new observaations and their 10 features}

\item{b}{a vector of coefficients obtained from the \code{topten} function}
}
\value{
a numeric vector containing the predicted values
}
\description{
This function takes a set of coeffifients produced by the \code{topten}
function and makes a prediction for each of the values provided in the
input 'X' matrix
}

