% Generated by roxygen2 (4.0.1): do not edit by hand
\name{top10}
\alias{top10}
\title{Building a Model with Top Ten Features}
\usage{
top10(x, y)
}
\arguments{
\item{x}{an [n x p] matrix of n observations an p predictors}

\item{y}{a vector of length n representing the response}
}
\value{
a vector of coefficients from the final fitted model with the top 10 features
}
\description{
This functions develops a prediction algorithm base on the top 10 features in X that are
the most predictive of 'y'
}
\details{
1) We'll fit a univariate regression model of the response on each of the predictors (one model for each predictor)
2) Every predictor is going to have a p-value associated with the model we fit
3) We'll sort those p-values from smallest to largest, taking the top 10.
4) Abstract a new dataset using only those predictors, and then fit a new model
5) Fit a new model using y ~ the 10 new predictors
6) Return the new model's coefficients
}
\author{
Brian Sipple
}
\seealso{
\code{lm}
}

