% Generated by roxygen2 (4.0.2): do not edit by hand
\name{top10predict}
\alias{top10predict}
\title{Building Model with top 10 feature}
\usage{
top10predict(x, y)
}
\arguments{
\item{x}{a n X p matrix of n observation and p predictors}

\item{y}{is a vector of length n respresenting the response}
}
\value{
a vector of coefficients from the fitted model with top 10 features
}
\description{
This function develops a prediction algorithm based on top 10 features
in "x" that are most predictive of "y"
}
\details{
This function run univariate  regression of y on each predictor in x and
calulate a p-values indicating  the significance  of the association. The
final set of 10 predictors is taken from the 10 smallest p-values
}
\author{
Abhishek Kapoor
}
\seealso{
\code {lm}
}

