% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict10}
\alias{predict10}
\title{Predicting with top 10 features}
\usage{
predict10(X, b)
}
\arguments{
\item{X}{a n x 10 matrix containing n new observations}

\item{b}{a vector of coefficients obtained from the \code{top10predict} function}
}
\value{
a numeric vector containing predicited values
}
\description{
This fuction takes a set of coefficients produced by the \code{top10predict}
function and makes prediction for each value provided in the input "X" matrix.
}

