% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{predict10}
\alias{predict10}
\title{Prediction with top 10 featurres
This function takes a set of coefficients produced by the \code{top10}}
\usage{
predict10(X, b)
}
\arguments{
\item{X}{a n x 10 matrix containing a new observations}
}
\value{
a numeric vector containing the predicted value
}
\description{
Prediction with top 10 featurres
This function takes a set of coefficients produced by the \code{top10}
}

