% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{top10}
\alias{top10}
\title{Building a model with top ten features}
\usage{
top10(x, y)
}
\arguments{
\item{x}{a n x p maxtri of n observation and p predictors}

\item{y}{a vector of lenght representing the response}
}
\value{
a vector of coefficeint from final model with top 10 features.
}
\description{
This function develops a prediction algorithm based on the top 10 features
in 'x' that are most predictive of 'y'.
}
\details{
This function runs a univariate regression of y on each predction in X and
calcuates a p-value indicating significance of the association. The final set of
10 predictors is taken from the 10 smallest p-vales.
}
\author{
Satyendra Raj Pandey
}
\seealso{
\code{lm}
}

