% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{top10}
\alias{top10}
\title{Building a Model with Top Ten Features}
\usage{
top10(x, y)
}
\arguments{
\item{x}{an n * p matrix of n observations and p predictors}

\item{y}{a vector of length n representing the response}
}
\value{
a vector of the coefficients from the final fitted model with
top ten features
}
\description{
This function develops a prediciton algorithm based on the top 10 features
in "x" that are most predictive of "y".
}
\details{
This function runs a univariate regression of y on each predictor in x and
calculates a p-value indicating the significance of the model with each
predictors. The final set of twn predictors are taken from the top ten
smallest p-values.
}
\seealso{
\code{lm}
}
\author{
Peter Siw
}
